requirements.txt
streamlit
pandas
folium
streamlit-folium
pydeck
numpy
plotly
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud
google-cloud-core
google-cloud-vision
google-api-core
google-generativeai
google-ai-generativelanguage
